+++ python/cudf/cudf/tests/test_dataframe.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2018-2025, NVIDIA CORPORATION.
 
 import array as arr
-import io
 import operator
 import textwrap
 from copy import copy
@@ -2343,147 +2342,6 @@ def test_df_sr_binop_col_order(op):
     assert_eq(expect, got)
 
 
-@pytest.mark.parametrize("set_index", [None, "A", "C", "D"])
-@pytest.mark.parametrize("index", [True, False])
-@pytest.mark.parametrize("deep", [True, False])
-def test_memory_usage(deep, index, set_index):
-    # Testing numerical/datetime by comparing with pandas
-    # (string and categorical columns will be different)
-    rows = 100
-    df = pd.DataFrame(
-        {
-            "A": np.arange(rows, dtype="int64"),
-            "B": np.arange(rows, dtype="int32"),
-            "C": np.arange(rows, dtype="float64"),
-        }
-    )
-    df["D"] = pd.to_datetime(df.A)
-    if set_index:
-        df = df.set_index(set_index)
-
-    gdf = cudf.from_pandas(df)
-
-    if index and set_index is None:
-        # Special Case: Assume RangeIndex size == 0
-        with expect_warning_if(deep, UserWarning):
-            assert gdf.index.memory_usage(deep=deep) == 0
-
-    else:
-        # Check for Series only
-        assert df["B"].memory_usage(index=index, deep=deep) == gdf[
-            "B"
-        ].memory_usage(index=index, deep=deep)
-
-        # Check for entire DataFrame
-        assert_eq(
-            df.memory_usage(index=index, deep=deep).sort_index(),
-            gdf.memory_usage(index=index, deep=deep).sort_index(),
-        )
-
-
-@pytest_xfail
-def test_memory_usage_string():
-    rows = 100
-    rng = np.random.default_rng(seed=0)
-    df = pd.DataFrame(
-        {
-            "A": np.arange(rows, dtype="int32"),
-            "B": rng.choice(["apple", "banana", "orange"], rows),
-        }
-    )
-    gdf = cudf.from_pandas(df)
-
-    # Check deep=False (should match pandas)
-    assert gdf.B.memory_usage(deep=False, index=False) == df.B.memory_usage(
-        deep=False, index=False
-    )
-
-    # Check string column
-    assert gdf.B.memory_usage(deep=True, index=False) == df.B.memory_usage(
-        deep=True, index=False
-    )
-
-    # Check string index
-    assert gdf.set_index("B").index.memory_usage(
-        deep=True
-    ) == df.B.memory_usage(deep=True, index=False)
-
-
-def test_memory_usage_cat():
-    rows = 100
-    rng = np.random.default_rng(seed=0)
-    df = pd.DataFrame(
-        {
-            "A": np.arange(rows, dtype="int32"),
-            "B": rng.choice(["apple", "banana", "orange"], rows),
-        }
-    )
-    df["B"] = df.B.astype("category")
-    gdf = cudf.from_pandas(df)
-
-    expected = (
-        gdf.B._column.categories.memory_usage
-        + gdf.B._column.codes.memory_usage
-    )
-
-    # Check cat column
-    assert gdf.B.memory_usage(deep=True, index=False) == expected
-
-    # Check cat index
-    assert gdf.set_index("B").index.memory_usage(deep=True) == expected
-
-
-def test_memory_usage_list():
-    df = cudf.DataFrame({"A": [[0, 1, 2, 3], [4, 5, 6], [7, 8], [9]]})
-    expected = (
-        df.A._column.offsets.memory_usage + df.A._column.elements.memory_usage
-    )
-    assert expected == df.A.memory_usage()
-
-
-def test_memory_usage_multi():
-    # We need to sample without replacement to guarantee that the size of the
-    # levels are always the same.
-    rng = np.random.default_rng(seed=0)
-    rows = 10
-    df = pd.DataFrame(
-        {
-            "A": np.arange(rows, dtype="int32"),
-            "B": rng.choice(
-                np.arange(rows, dtype="int64"), rows, replace=False
-            ),
-            "C": rng.choice(
-                np.arange(rows, dtype="float64"), rows, replace=False
-            ),
-        }
-    ).set_index(["B", "C"])
-    gdf = cudf.from_pandas(df)
-    # Assume MultiIndex memory footprint is just that
-    # of the underlying columns, levels, and codes
-    expect = rows * 16  # Source Columns
-    expect += rows * 16  # Codes
-    expect += rows * 8  # Level 0
-    expect += rows * 8  # Level 1
-
-    assert expect == gdf.index.memory_usage(deep=True)
-
-
-@pytest.mark.parametrize("index", [False, True])
-def test_memory_usage_index_preserve_types(index):
-    data = [[1, 2, 3]]
-    columns = pd.Index(np.array([1, 2, 3], dtype=np.int8), name="a")
-    result = (
-        cudf.DataFrame(data, columns=columns).memory_usage(index=index).index
-    )
-    expected = (
-        pd.DataFrame(data, columns=columns).memory_usage(index=index).index
-    )
-    if index:
-        # pandas returns an Index[object] with int and string elements
-        expected = expected.astype(str)
-    assert_eq(result, expected)
-
-
 @pytest.mark.parametrize(
     "list_input",
     [
@@ -2685,318 +2543,3 @@ def test_dataframe_assign_scalar_with_scalar_cols(col_data, assign_val):
     gdf["b"] = assign_val
 
     assert_eq(pdf, gdf)
-
-
-def test_dataframe_info_basic():
-    buffer = io.StringIO()
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    Index: 10 entries, a to 1111
-    Data columns (total 10 columns):
-     #   Column  Non-Null Count  Dtype
-    ---  ------  --------------  -----
-     0   0       10 non-null     float64
-     1   1       10 non-null     float64
-     2   2       10 non-null     float64
-     3   3       10 non-null     float64
-     4   4       10 non-null     float64
-     5   5       10 non-null     float64
-     6   6       10 non-null     float64
-     7   7       10 non-null     float64
-     8   8       10 non-null     float64
-     9   9       10 non-null     float64
-    dtypes: float64(10)
-    memory usage: 859.0+ bytes
-    """
-    )
-    rng = np.random.default_rng(seed=0)
-    df = pd.DataFrame(
-        rng.standard_normal(size=(10, 10)),
-        index=["a", "2", "3", "4", "5", "6", "7", "8", "100", "1111"],
-    )
-    cudf.from_pandas(df).info(buf=buffer, verbose=True)
-    s = buffer.getvalue()
-    assert str_cmp == s
-
-
-def test_dataframe_info_verbose_mem_usage():
-    buffer = io.StringIO()
-    df = pd.DataFrame({"a": [1, 2, 3], "b": ["safdas", "assa", "asdasd"]})
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 3 entries, 0 to 2
-    Data columns (total 2 columns):
-     #   Column  Non-Null Count  Dtype
-    ---  ------  --------------  -----
-     0   a       3 non-null      int64
-     1   b       3 non-null      object
-    dtypes: int64(1), object(1)
-    memory usage: 56.0+ bytes
-    """
-    )
-    cudf.from_pandas(df).info(buf=buffer, verbose=True)
-    s = buffer.getvalue()
-    assert str_cmp == s
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 3 entries, 0 to 2
-    Columns: 2 entries, a to b
-    dtypes: int64(1), object(1)
-    memory usage: 56.0+ bytes
-    """
-    )
-    cudf.from_pandas(df).info(buf=buffer, verbose=False)
-    s = buffer.getvalue()
-    assert str_cmp == s
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    df = pd.DataFrame(
-        {"a": [1, 2, 3], "b": ["safdas", "assa", "asdasd"]},
-        index=["sdfdsf", "sdfsdfds", "dsfdf"],
-    )
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    Index: 3 entries, sdfdsf to dsfdf
-    Data columns (total 2 columns):
-     #   Column  Non-Null Count  Dtype
-    ---  ------  --------------  -----
-     0   a       3 non-null      int64
-     1   b       3 non-null      object
-    dtypes: int64(1), object(1)
-    memory usage: 91.0 bytes
-    """
-    )
-    cudf.from_pandas(df).info(buf=buffer, verbose=True, memory_usage="deep")
-    s = buffer.getvalue()
-    assert str_cmp == s
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    int_values = [1, 2, 3, 4, 5]
-    text_values = ["alpha", "beta", "gamma", "delta", "epsilon"]
-    float_values = [0.0, 0.25, 0.5, 0.75, 1.0]
-
-    df = cudf.DataFrame(
-        {
-            "int_col": int_values,
-            "text_col": text_values,
-            "float_col": float_values,
-        }
-    )
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 5 entries, 0 to 4
-    Data columns (total 3 columns):
-     #   Column     Non-Null Count  Dtype
-    ---  ------     --------------  -----
-     0   int_col    5 non-null      int64
-     1   text_col   5 non-null      object
-     2   float_col  5 non-null      float64
-    dtypes: float64(1), int64(1), object(1)
-    memory usage: 130.0 bytes
-    """
-    )
-    df.info(buf=buffer, verbose=True, memory_usage="deep")
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-
-def test_dataframe_info_null_counts():
-    int_values = [1, 2, 3, 4, 5]
-    text_values = ["alpha", "beta", "gamma", "delta", "epsilon"]
-    float_values = [0.0, 0.25, 0.5, 0.75, 1.0]
-
-    df = cudf.DataFrame(
-        {
-            "int_col": int_values,
-            "text_col": text_values,
-            "float_col": float_values,
-        }
-    )
-    buffer = io.StringIO()
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 5 entries, 0 to 4
-    Data columns (total 3 columns):
-     #   Column     Dtype
-    ---  ------     -----
-     0   int_col    int64
-     1   text_col   object
-     2   float_col  float64
-    dtypes: float64(1), int64(1), object(1)
-    memory usage: 130.0+ bytes
-    """
-    )
-    df.info(buf=buffer, verbose=True, null_counts=False)
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    df.info(buf=buffer, verbose=True, max_cols=0)
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    df = cudf.DataFrame()
-
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 0 entries
-    Empty DataFrame"""
-    )
-    df.info(buf=buffer, verbose=True)
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    df = cudf.DataFrame(
-        {
-            "a": [1, 2, 3, None, 10, 11, 12, None],
-            "b": ["a", "b", "c", "sd", "sdf", "sd", None, None],
-        }
-    )
-
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 8 entries, 0 to 7
-    Data columns (total 2 columns):
-     #   Column  Dtype
-    ---  ------  -----
-     0   a       int64
-     1   b       object
-    dtypes: int64(1), object(1)
-    memory usage: 238.0+ bytes
-    """
-    )
-    pd.options.display.max_info_rows = 2
-    df.info(buf=buffer, max_cols=2, null_counts=None)
-    pd.reset_option("display.max_info_rows")
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    str_cmp = textwrap.dedent(
-        """\
-    <class 'cudf.core.dataframe.DataFrame'>
-    RangeIndex: 8 entries, 0 to 7
-    Data columns (total 2 columns):
-     #   Column  Non-Null Count  Dtype
-    ---  ------  --------------  -----
-     0   a       6 non-null      int64
-     1   b       6 non-null      object
-    dtypes: int64(1), object(1)
-    memory usage: 238.0+ bytes
-    """
-    )
-
-    df.info(buf=buffer, max_cols=2, null_counts=None)
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-    buffer.truncate(0)
-    buffer.seek(0)
-
-    df.info(buf=buffer, null_counts=True)
-    actual_string = buffer.getvalue()
-    assert str_cmp == actual_string
-
-
-@pytest.mark.parametrize(
-    "df",
-    [
-        pd.DataFrame({"a": [1, 2, 3, 4, 5, 10, 11, 12, 33, 55, 19]}),
-        pd.DataFrame(
-            {
-                "one": [1, 2, 3, 4, 5, 10],
-                "two": ["abc", "def", "ghi", "xyz", "pqr", "abc"],
-            }
-        ),
-        pd.DataFrame(
-            {
-                "one": [1, 2, 3, 4, 5, 10],
-                "two": ["abc", "def", "ghi", "xyz", "pqr", "abc"],
-            },
-            index=[10, 20, 30, 40, 50, 60],
-        ),
-        pd.DataFrame(
-            {
-                "one": [1, 2, 3, 4, 5, 10],
-                "two": ["abc", "def", "ghi", "xyz", "pqr", "abc"],
-            },
-            index=["a", "b", "c", "d", "e", "f"],
-        ),
-        pd.DataFrame(index=["a", "b", "c", "d", "e", "f"]),
-        pd.DataFrame(columns=["a", "b", "c", "d", "e", "f"]),
-        pd.DataFrame(index=[10, 11, 12]),
-        pd.DataFrame(columns=[10, 11, 12]),
-        pd.DataFrame(),
-        pd.DataFrame({"one": [], "two": []}),
-        pd.DataFrame({2: [], 1: []}),
-        pd.DataFrame(
-            {
-                0: [1, 2, 3, 4, 5, 10],
-                1: ["abc", "def", "ghi", "xyz", "pqr", "abc"],
-                100: ["a", "b", "b", "x", "z", "a"],
-            },
-            index=[10, 20, 30, 40, 50, 60],
-        ),
-    ],
-)
-def test_dataframe_keys(df):
-    gdf = cudf.from_pandas(df)
-
-    assert_eq(
-        df.keys(),
-        gdf.keys(),
-    )
-
-
-@pytest.mark.parametrize(
-    "ps",
-    [
-        pd.Series([1, 2, 3, 4, 5, 10, 11, 12, 33, 55, 19]),
-        pd.Series(["abc", "def", "ghi", "xyz", "pqr", "abc"]),
-        pd.Series(
-            [1, 2, 3, 4, 5, 10],
-            index=["abc", "def", "ghi", "xyz", "pqr", "abc"],
-        ),
-        pd.Series(
-            ["abc", "def", "ghi", "xyz", "pqr", "abc"],
-            index=[1, 2, 3, 4, 5, 10],
-        ),
-        pd.Series(index=["a", "b", "c", "d", "e", "f"], dtype="float64"),
-        pd.Series(index=[10, 11, 12], dtype="float64"),
-        pd.Series(dtype="float64"),
-        pd.Series([], dtype="float64"),
-    ],
-)
-def test_series_keys(ps):
-    gds = cudf.from_pandas(ps)
-
-    assert_eq(ps.keys(), gds.keys())
